% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reactable_columns.R
\name{col_badge}
\alias{col_badge}
\title{Reactable badge column}
\usage{
col_badge(
  name,
  tooltip = NULL,
  colors,
  ...,
  min_width = 75,
  border_right = NULL
)
}
\arguments{
\item{name}{\code{chr} Display name of the column to be used as the header.}

\item{tooltip}{\code{chr} Tooltip text to be added to the header. Default is NULL.}

\item{colors}{\code{list} Named list of CSS appropriate strings indicating the badge colors. Keys
should be the data values associated with each color.}

\item{...}{Additional arguments to be passed to \link[reactable:reactable]{reactable::reactable}.}

\item{min_width}{\code{int} Minimum width for the column.}

\item{border_right}{\code{bool / chr} Right side border to be added. Default is NULL.}
}
\description{
Creates a \link[reactable:colDef]{reactable::colDef} object with colored ovals in the background. Useful
for columns of categorical data.
}
