% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reactable_columns.R
\name{col_base}
\alias{col_base}
\title{Reactable base column}
\usage{
col_base(
  name,
  tooltip = NULL,
  color = "#000000",
  digits = 0,
  percent = FALSE,
  ...,
  min_width = 75,
  border_right = NULL
)
}
\arguments{
\item{name}{\code{chr} Display name of the column to be used as the header.}

\item{tooltip}{\code{chr} Tooltip text to be added to the header. Default is NULL.}

\item{color}{\code{chr} CSS appropriate string indicating the font color.}

\item{digits}{\code{int} Number of decimal digits to display. Default is 0.}

\item{percent}{\code{bool} Whether to format the column as a percent. Default is FALSE.}

\item{...}{Additional arguments to be passed to \link[reactable:reactable]{reactable::reactable}.}

\item{min_width}{\code{int} Minimum width for the column.}

\item{border_right}{\code{bool / chr} Right side border to be added. Default is NULL.}
}
\description{
Creates a \link[reactable:colDef]{reactable::colDef} object.
}
