% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reactable_columns.R
\name{col_sparkline}
\alias{col_sparkline}
\title{Reactable sparkline column}
\usage{
col_sparkline(
  data,
  name = "Spark",
  tooltip = NULL,
  color = "#c5c5c5",
  ...,
  border_right = NULL,
  max_width = 60
)
}
\arguments{
\item{data}{\code{vct} Vector of values to be plotted in the sparkline.}

\item{name}{\code{chr} Display name of the column to be used as the header. Default is \code{'Spark'}.}

\item{tooltip}{\code{chr} Tooltip text to be added to the header. Default is NULL.}

\item{color}{\code{chr} CSS appropriate string indicating the sparkline bar color.}

\item{...}{Additional arguments to be passed to \link[reactable:reactable]{reactable::reactable}.}

\item{border_right}{\code{bool / chr} Right side border to be added. Default is NULL.}

\item{max_width}{\code{int} Maximum width for the column.}
}
\description{
Creates a \link[reactable:colDef]{reactable::colDef} object with sparklines.
}
