% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_grid_titles.R
\name{e_grid_titles}
\alias{e_grid_titles}
\title{Titles for Echarts grid}
\usage{
e_grid_titles(e, titles, top_space = 2)
}
\arguments{
\item{e}{\code{echart} Chart to be modified.}

\item{titles}{\code{vector / list} Names to be used. \strong{Note}: \code{\link[=e_grid_titles]{e_grid_titles()}} will use the order
in \code{titles} to assign titles along the grid (left to right, top to bottom). So, if you made a
facet plot using a charcter grouping variable, you will probably need to sort \code{titles}
alphabetically for the order to match the grid.}

\item{top_space}{\code{int} How far above the plot the title should be placed. The units of
\code{top_space} (percent vs pixels) will be determined automatically to match the units used in
the specification of the grid. Default value is 2.}
}
\description{
Adds titles to echart subplots built using a grid (for example with
\code{\link[echarts4r:e_facet]{echarts4r::e_facet()}}. \strong{Note}: This function will work with grids specified entirely in
percent or entirely in pixels only. It will not work for grids using a mix of pixels and percent.
TODO: need to make this work for pixels.
}
\examples{
groups <- 6
observations <- 10
df <- data.frame(x = rep(1:observations,
                         times = groups),
                 y = runif(groups * observations, 0, 1),
                 group = rep(paste('Group', 1:6), each = observations))

df \%>\%
  dplyr::group_by(group) \%>\%
  echarts4r::e_charts(x) \%>\%
  echarts4r::e_bar(y,
                   name = 'Outcome') \%>\%
  echarts4r::e_facet(rows = 2,
                     cols = 3,
                     legend_space = 5) \%>\%
  e_grid_titles(titles = unique(df$group))
}
