% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_map.R
\name{make_leaflet_labels}
\alias{make_leaflet_labels}
\title{Make leaflet tooltip labels}
\usage{
make_leaflet_labels(
  data,
  id_label = 1,
  info_labels,
  decimals = 0,
  percent = c(),
  evolution = c(),
  flip_colors = c()
)
}
\arguments{
\item{data}{\verb{spatial polygon dataframe} Merged map data.}

\item{id_label}{\code{chr} Column name in \code{df} to use as the shape name in the tooltip.}

\item{info_labels}{\code{list} Named list corresponding to the \code{df} columns (keys) and desired label
text (values) to appear in the tooltip. By default, the list \code{x} will be used.}

\item{decimals}{\code{int} Number of decimal places to be shown. Passed to \code{\link[tinker:french_format]{tinker::french_format()}}.
Default is 0.}

\item{percent}{\verb{chr vector} Vector listing the names of variables in \code{info_labels} to be
displayed as percents. Passed to \code{\link[=french_format]{french_format()}}. Note that \code{\link[=french_format]{french_format()}} assumes the
values are \emph{proportions} not \emph{percents} (ie, to get 43\%, you must input 0.43 not 43). Default
is an empty list.}

\item{evolution}{\verb{chr vector} Vector listing the names of variables in \code{info_labels} to have
their evolution included in the tooltip. \strong{Warning}, \code{\link[=make_leaflet_labels]{make_leaflet_labels()}} assumes that
these evolutions are already in \code{data} with column names conforming to the format {name}_evo.
Default is an empty list.}

\item{flip_colors}{\verb{chr vector} Vector listing the names of variables in \code{info_labels} where
the default evolution colors should be inverted. Passed to \code{\link[tinker:up_down_format]{tinker::up_down_format()}}. Default
is an empty list.}
}
\value{
\code{list}
}
\description{
Build the html tooltip text.
}
