% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reactable.R
\name{render_reactable}
\alias{render_reactable}
\title{Render Reactable}
\usage{
render_reactable(
  df,
  columns = NULL,
  column_groups = NULL,
  digits = 0,
  elementId = "my-select",
  fullWidth = FALSE,
  searchable = TRUE,
  page_size = 25,
  sort_by = NULL,
  sort_order = "desc",
  ...
)
}
\arguments{
\item{df}{\code{dataframe} Data to be tablized.}

\item{columns}{\code{list} Column definitions, should follow the standard reactable format. Default is
NULL.}

\item{column_groups}{\code{list} Convenience arggument to simplify setting column groups. Should be a
named list where keys are the group names and values are vectors of the column names within
the corresponding group. Default is NULL.}

\item{digits}{\code{int} Default number of digits to show for numeric columns. Default is 0.}

\item{elementId}{\code{chr} Table id, used primarily for JS functions. Default is \code{'my-select'}.}

\item{fullWidth}{\code{bool} Whether tables should be expanded to fill a window by default. Default is
FALSE.}

\item{searchable}{\code{bool} Whether tables should be searchable by default. Default is TRUE.}

\item{page_size}{\code{int} Number of observations to show per page. Default is 25.}

\item{sort_by}{\code{chr} Name of column to sort the table by. Default is NULL.}

\item{sort_order}{\code{chr} Whether to sort the \code{sort_by} column in ascending or descending order.
Default is \code{'desc'}.}

\item{...}{Additional arguments to be passed to \link[reactable:reactable]{reactable::reactable}.}
}
\description{
Function to create a general reactable table with sensible defaults.
}
