% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reactable_styles.R
\name{render_reactable_bar}
\alias{render_reactable_bar}
\title{Reactable bar cell}
\usage{
render_reactable_bar(
  width,
  fill = "#e6e6e6",
  color = "#000000",
  font_weight = "bold",
  border_right = NULL
)
}
\arguments{
\item{width}{\code{num} Proportion (not percent) of cell to fill with the bar (width wise).}

\item{fill}{\code{chr} CSS appropriate string indicating bar fill.}

\item{color}{\code{chr} CSS appropriate sting indicating text color.}

\item{font_weight}{\code{chr} Font weight.}

\item{border_right}{\code{bool / chr} Border to be used on the right side of the cell. If \code{TRUE} a
default border style of a thin grey line is used. The user can also pass a CSS style string.}
}
\description{
Adds bar styling to a reactable cell through embedded CSS.
}
