% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reactable_styles.R
\name{render_reactable_border}
\alias{render_reactable_border}
\title{Reactable cell border}
\usage{
render_reactable_border(border = NULL)
}
\arguments{
\item{border}{\code{bool / chr} A user can either pass \code{TRUE} to add a default border (thin and grey)
or they can pass their own CSS style string. Anything else will result in no border. Default
is NULL.}
}
\description{
Convenience function to add a cell border to a reactable.
}
