% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reactable_styles.R
\name{render_reactable_header}
\alias{render_reactable_header}
\title{Reactable headers (with tooltips)}
\usage{
render_reactable_header(name, tooltip = NULL)
}
\arguments{
\item{name}{\code{chr} Column name.}

\item{tooltip}{\code{chr} Tooltip to be added to the header text. Default is NULL (no tooltip).}
}
\description{
Convenience function to facilitate adding tooltips to reactable headers.
}
